cargo: bump rust-version and dependencies
Get rid of parking_lot in Tokio as dicussed in 986d668209. Because we
need to specify features instead of the "full" shortcut, the single
threaded runtime has also been remove.

Furthermore, I think its best we specify the full version of the
dependencies. Tools like cargo-outdated will help with the maintainance
part and should be used regardless anyway.